requirements.txt
loguru
pip
python-dotenv
ruff
tqdm
typer
pandas
matplotlib
seaborn
scikit-learn
shap
tensorflow
kagglehub 

captum
-e .